+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All Notable changes to `Csv` will be documented in this file
 - `TabulatDataReader::firstOrFailMatching`
 - `FragmentFinder` to implement [RFC7111](https://www.rfc-editor.org/rfc/rfc7111)
 - `ResultSet::fromRecords`
+- `Stream::setMaxLineLen`
+- `Stream::getMaxLineLen`
 
 ### Deprecated
 
@@ -26,6 +28,7 @@ It's usage will trigger a `E_USER_DEPRECATED` call.
 - `ResultSet` constructor now allows the records to be an `array`.
 - to the internal `Stream` object it will throw a `RuntimeException` if the rewind action fails
 - if calls to `fseek` fails (returns `-1` ) a new `RuntimeException` will be thrown too.
+- `Stream` can iterate and return the full line respecting `SplFielObject` flags. Previously it only returned the CSV records.
 
 ### Removed